backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings
celery==5.4.0
redis==5.0.3
aio_pika==9.4.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.0
python-dateutil==2.9.0.post0

# Real-time Emotion Processing Dependencies
websockets==12.0
numpy==1.24.3
aiohttp==3.9.5
scikit-learn==1.3.0

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.1